renaming and changing to enums
