+++ Apps/AdminWebClient/src/ClientApp/src/components/core/modals/EmailModal.vue
@@ -64,7 +64,7 @@
 <script lang="ts">
 import { Component, Vue, Watch, Emit, Prop } from "vue-property-decorator";
 import container from "@/plugins/inversify.config";
-import Communication from "@/models/communication";
+import Communication from "@/models/adminCommunication";
 import { ResultType } from "@/constants/resulttype";
 import moment from "moment";
 import {
